feat: Update convert flac to mp3

+++ convert_flac_to_mp3.py
@@ -4,6 +4,51 @@ from pydub import AudioSegment
 import multiprocessing
 from functools import partial
 import random
+from mutagen.flac import FLAC
+from mutagen.mp3 import MP3
+from mutagen.id3 import ID3, TIT2, TPE1, TALB, TCON, TDRC, APIC
+import sys
+
+def handle_metadata(flac_path, mp3_path):
+    flac_metadata = FLAC(flac_path)
+
+    # Create an MP3 file object to add metadata
+    mp3_metadata = MP3(mp3_path, ID3=ID3)
+    
+    # Add ID3 tag if it doesn't exist
+    try:
+        mp3_metadata.add_tags()
+    except:
+        pass
+
+    # Transfer metadata from FLAC to MP3
+    for tag in flac_metadata.keys():
+        if tag == "title":
+            mp3_metadata.tags.add(TIT2(encoding=3, text=flac_metadata[tag]))
+        elif tag == "artist":
+            mp3_metadata.tags.add(TPE1(encoding=3, text=flac_metadata[tag]))
+        elif tag == "album":
+            mp3_metadata.tags.add(TALB(encoding=3, text=flac_metadata[tag]))
+        elif tag == "genre":
+            mp3_metadata.tags.add(TCON(encoding=3, text=flac_metadata[tag]))
+        elif tag == "date":
+            mp3_metadata.tags.add(TDRC(encoding=3, text=flac_metadata[tag]))
+
+    for picture in flac_metadata.pictures:
+        mp3_metadata.tags.add(
+            APIC(
+                encoding=3,         # 3 is for utf-8
+                mime=picture.mime,  # image/jpeg or image/png
+                type=3,             # 3 is for the cover (front) image
+                desc=u'Cover',
+                data=picture.data
+            )
+        )
+
+
+    # Save the MP3 file with metadata
+    mp3_metadata.save()
+
 
 def convert_flac_to_mp3(files, src_folder, dst_folder, bitrate="320k"):
     # Ensure the destination folder exists
@@ -18,20 +63,29 @@ def convert_flac_to_mp3(files, src_folder, dst_folder, bitrate="320k"):
         if filename.endswith(".flac"):
             try:
                 dst_file = os.path.join(dst_folder, f"{os.path.splitext(filename)[0]}.mp3")
+                if os.path.isfile(dst_file):
+                    print(f"EXIST: {filename}")
+                    continue
 
                 # Load the FLAC file
                 audio = AudioSegment.from_file(src_file, format="flac")
                 # Export as MP3 with the specified bitrate
-                audio.export(dst_file, format="mp3", bitrate=bitrate)
+                audio.export(dst_file, format="mp3", bitrate=bitrate, parameters=["-ac", "2", "-ar", "48000"])
+                handle_metadata(src_file, dst_file)
 
                 cnt_convert += 1
                 print(f"SUCCESS - {os.getpid()}: {src_file} to {dst_file}")
             except BaseException as e:
-                print(f"ERROR - {os.getpid()}: {e}")
                 print(f"FAILED - {os.getpid()}: {src_file}")
+                print(f"ERROR - {os.getpid()}: {e}")
                 errs.append(filename)
 
         else:
+            dst_file = os.path.join(dst_folder, filename)
+            if os.path.isfile(dst_file):
+                    print(f"EXIST: {filename}")
+                    continue
+            
             shutil.copy2(src_file, dst_folder)
             cnt_cp += 1
             print(f"COPY - {os.getpid()}: {src_file} to {dst_folder}")
@@ -39,7 +93,8 @@ def convert_flac_to_mp3(files, src_folder, dst_folder, bitrate="320k"):
     print(f"SUCCESS: {cnt_convert}")
     print(f"COPY: {cnt_cp}")
     print(f"Total: {cnt_convert + cnt_cp}")
-    print(f"FAILED - {os.getpid()}: {len(errs), errs}")
+    print(f"FAIL: {os.getpid()}: {len(errs), errs}")
+    sys.stdout.flush()
 
 
 def divide_list(lst, n):
@@ -56,7 +111,7 @@ def divide_list(lst, n):
 # Set your source and destination folders
 src_folder = "../files/Rocks"
 dst_folder = "../files/Rocks_compressed"
-NUM_PROCESS = 10
+NUM_PROCESS = 12
 
 if __name__ == "__main__":
     files = os.listdir(src_folder)
